+++ metaServer/src/main/java/com/ksyun/campus/metaserver/pojo/DataServerMsg.java
@@ -9,6 +9,6 @@ public class DataServerMsg {
     private String host;
     private int port;
     private int fileTotal;
-    private int capacity; //单位：字节B
-    private int useCapacity;
+    private long capacity; //单位：字节B
+    private long useCapacity;
 }
